Fixed #198 (at least it's not so ugly now).
However, we don't know how to perform a conditional rendering
in both cases when we already output something and, the second
case, when page is empty. This leads to invalid HTML, and this
problem should be resolved separately. As a solution we can
have a flag that controls output state (except headers), so each
auth form can control what to do: either render the complete body
or behave like a widget.