Fixed bug #256 (Cheap user deletion: fine manual).
The problem rooted in the attempt to read old-style KV-stored values
using the brand new mechanism. Of course, it failed with
a little bit unpredictable results. As no basic error checking
was provided, all edited users really gone to Hell after
a couple of touches.

To prevent this odd behaviour, a SPIKE was added to use old-style
KV API functions (and a old-style template was added to handle
old-style data).