Fixed #366 (Custom rewrite rules).
In a presence of the file $content_dir/cms/custom-htaccess,
it will be included into autogenerated .htaccess in the site root.
It should be syntactically correct, otherwise you'll get a HTTP 500
server misconfiguration problem.

Additionally, a bug was fixed: appending content to .htaccess
instead of rewriting it (so .htaccess can grow ad infinitum).