No more elementsdir, engine_dir used everywhere.
Eliminated elementsdir hell, that was a synonym of
'$engine_dir' global variable. Moreover, new technology
installer used the "engine_dir" key in $SETTINGS, so it
was only the coincidence that all this worked: if somebody
renamed engine directory, a half of code would fail.

Still don't understand how the new installer was written
without this massive refactoring.