Fixed wrong content destination installation.
When performing installation twice, second time the content
folder was copied *inside* the content folder. That is because
of 'cp' behaviour:

cp -a src dest/qqq

behaviour depends on existance of 'dest/qqq' folder. When it does
not exist, 'src' will be copied to 'dest/' as 'qqq', but if 'dest/qqq'
already exists, 'src' will be copied inside 'dest/qqq', so it will be
'dest/qqq/src'.

So we should remove target directory before copying. Also this ensures
we have clean content without old files.