Apply a patch from pbagnall@: trailing whitespace patch.
Fixed bug where "abc def" and "abc def ghi" when using the
wordDelimiters would result in "c4d3i7:def ghi" rather than "c7i4: ghi".
The issue was that whitespace was being treated as being part of a word,
so the lack of trailing whitespace after "def" meant that it didn't
match "def " and therefore wasn't recognised as being a valid match. The
solution is to create separate word fragments and whitespace fragments.

https://github.com/thomasbachem/PHP-FineDiff/commit/db098a6cd519d3592a54b00f824eb7be0eec5313

mvel@: This patch was not verified by unittests but seems appropriate.