Do not spoil cache on installer run.
There was a wonderful bug in our installer system. When 'install.php'
is detected in index.php, it is not sufficient to emit 'Location' header,
but also necessary *NOT TO RUN* all other 'index.php' code
that will run XCMS compiler.

This bug lead to weird behavior because cache was trashed *BEFORE*
installer is able to perform all its actions (server code executed
first and only then redirection in browser to install.php occured).