Fixed local-install bug: show usage on clean installation.
When destination was completely erased by external forces
or installation is performed at the first time, there
will be no TEMP_DB, and DB_MODE will not be set at all
if run script w/o parameters. Without this fix we just print
usage, that's confusing.

So now the behavior is clear: when no options specified,
we try to create a DB backup, but if it does not exist,
we just copy all the content (including DB) from sample
directory.